Change user_id type from int to float in User model

+++ domain/entities/user.py
@@ -2,11 +2,14 @@
 from pydantic import BaseModel, EmailStr, Field
 from typing import Optional
 
+
 class User(BaseModel):
-    user_id: int
+    user_id: float
     username: str
     email: EmailStr
     password: str = Field(..., min_length=1)
+
+
 class UpdateUser(BaseModel):
     username: Optional[str]
     email: EmailStr
